New major version 3. Added PortAudio support

+++ src/fft/fft.h
@@ -7,4 +7,4 @@ static inline int check_pow2(size_t n) {
     return n > 0 && (n & (n - 1)) == 0;
 }
 
-int fft(size_t n, double complex samples[static n]);
+int fft(size_t n, float complex samples[static n]);
